Fix promotion create and logic and make it more structural
