+++ backend/package.json
@@ -12,6 +12,7 @@
   "author": "CSC309 Student, Winter 2025",
   "license": "ISC",
   "dependencies": {
+    "@faker-js/faker": "^9.6.0",
     "@prisma/client": "^6.5.0",
     "bcrypt": "^5.1.1",
     "cors": "^2.8.5",
